ai_engine/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
imbalanced-learn==0.12.3
shap==0.45.1
fairlearn==0.10.0
openai==1.35.3
google-cloud-storage==2.17.0
google-cloud-bigquery==3.25.0
google-cloud-aiplatform==1.60.0
vertexai==1.60.0
pydantic==2.7.4
python-dotenv==1.0.1
requests==2.32.3
joblib==1.4.2
scipy==1.13.1